fix(experiments): relocating experiment modal to the right location.
